Remove redundant copy of SpecializeToVersion type family.
