+++ libs/wire-api/wire-api.cabal
@@ -324,6 +324,7 @@ library
     , servant-client-core
     , servant-conduit
     , servant-multipart
+    , servant-multipart-api
     , servant-openapi3
     , servant-server
     , singletons
